% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id.R
\name{add_id}
\alias{add_id}
\title{Add ID Field}
\usage{
add_id(df, level)
}
\arguments{
\item{df}{data frame; An FMG table to add an ID field to.}

\item{level}{character; The FMG level to create the ID field from.}
}
\description{
Add ID Field
}
