% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_needed_packages.R
\name{install_needed_packages}
\alias{install_needed_packages}
\title{Install needed packages}
\usage{
install_needed_packages(need_pkgs, source = "no")
}
\arguments{
\item{need_pkgs}{A character vector of package names.}

\item{source}{character; Indicating if the packages should be 
compiled from source ("yes", "no").}
}
\value{
Installs the requested packages.
}
\description{
Tests if packages are installed and if not installs them.
}
