% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load needed packages}
\usage{
load_packages(need_pkgs)
}
\arguments{
\item{need_pkgs}{A character vector of package names.}
}
\value{
Loads the requested packages.
}
\description{
Loads specified packages.
}
\examples{
need_pkgs <- c("testthat", "fmgr")

}
