% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_info.R
\name{param_info}
\alias{param_info}
\title{Parameter information}
\usage{
param_info(param, param_name)
}
\arguments{
\item{param}{character; The parameter.}

\item{param_name}{character; The name of the parameter.}
}
\value{
A single-row data frame containing the characteristics of a parameter.
}
\description{
Calculates the characteristics of a parameter.
}
