% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_table.R
\name{param_table}
\alias{param_table}
\title{Parameters information table}
\usage{
param_table(param_list)
}
\arguments{
\item{param_list}{list; A list of named parameter values.}
}
\value{
A data frame of parameters and their characteristics.
}
\description{
Calculates a table of parameter characteristics for a set of 
parameters.
}
