% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_na.R
\name{set_na}
\alias{set_na}
\title{Convert missing data to NA}
\usage{
set_na(df)
}
\arguments{
\item{df}{data frame; A data frame.}
}
\value{
The data frame with the covered instances of missing data converted 
to R NA.
}
\description{
Convert the commonly experienced forms of missing data to NA 
for all fields in a data frame.
}
